+++ Changelog.md
@@ -10,6 +10,7 @@ Compiler Features:
 
 Bugfixes:
 * Assembly-Json: Fix assembly json export to store jump types of operations in `jumpType` field instead of `value`.
+* TypeChecker: Convert parameters of function type to how they would be called for ``abi.encodeCall``.